+++ CHANGELOG.md
@@ -59,6 +59,11 @@ All notable changes to the "todo-tools" extension will be documented in this fil
 - save project name to archive
 
 ## [0.7.5] - 2022-09-11
+### add
+- setDueRelative (which will now be the default) will change days relative to the current due date
+- assigned setDueRelative to ctrl+alt+period period
+- assigned setDueRelative(-1) to ctrl+alt+period comma
+
 ### fix
 - autoRun troubleshooting